requirements.txt
Flask==2.2.5
torch==1.13.1
torchvision==0.14.1
Pillow==9.5.0
numpy==1.23.5
timm  
